requirements.txt
graphiti-core>=0.9.5
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0 